[Johnny 2022/08/24 繳交前最後更新 API皆可使用 但無JWT驗證]

+++ database/migrations/2022_08_19_192853_create_payment_table.php
@@ -15,7 +15,7 @@ class CreatePaymentTable extends Migration
     {
         Schema::create('payments', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('total');
+            $table->decimal('total');
             $table->string('paymentstatus');
             $table->timestamps();
             $table->softDeletes();
